Stop/Limit writing to cassandra from collector when disk full
- separated watermark logic to separate lib. Used by
    - queue_task (existing)
    - disk-usage-percentage/pending-compaction-tasks (new)
- writing messages to collector is based on 3 parameters
    - disk_usage on db nodes
    - pending-compaction-tasks on db nodes
    - severity level of incoming messages
- we configure disk-usage-percentage/pending-compaction-tasks watermarks map to
  message severity-level on collector on startup
- this is configurable from collector config file or command line

- opserver fetches disk-usage-percentage and pending-compaction-tasks for all
  db nodes periodically from uve
- opservers sends max disk-usage-percentage and max pending-compaction-tasks
  amongst all db nodes to collector
- following values can be accessed via introspect
    - disk-usage-percentage                    - get/set on opserver/collector
    - pending-compaction-tasks                 - get/set on opserver/collector
    - derived disk-usage-percentage-drop-level - get on collector
    - derived pending-compaction-tasks-severity-level
                                               - get on collector
- Added automated UT for all above changes using the above introspect api

Closes-Bug:1604200
