+++ base/SConscript
@@ -64,6 +64,7 @@ libbase = env.Library('base',
                        'lifetime.cc',
                        'logging.cc',
                        'proto.cc',
+                       'watermark.cc',
                        task,
                        'task_annotations.cc',
                        'task_sandesh.cc',
